+++ examples/reown/src/config/appkit.ts
@@ -1,0 +1,8 @@
+export const projectId = import.meta.env.VITE_REOWN_APP_ID
+
+export const metadata = {
+  name: 'LIFI Widget',
+  description: 'LI.FI Widget AppKit Example',
+  url: 'https://li.fi',
+  icons: ['https://avatars.githubusercontent.com/u/85288935'],
+}