+++ packages/widget/src/pages/TransactionHistoryPage/TransactionHistoryPage.tsx
@@ -1,11 +1,14 @@
 import type { FullStatusData } from '@lifi/sdk'
-import { List } from '@mui/material'
+import { Box, List } from '@mui/material'
 import { useVirtualizer } from '@tanstack/react-virtual'
+import type React from 'react'
 import { useRef } from 'react'
 import { useTranslation } from 'react-i18next'
 import { PageContainer } from '../../components/PageContainer.js'
 import { useHeader } from '../../hooks/useHeader.js'
 import { useTransactionHistory } from '../../hooks/useTransactionHistory.js'
+
+import { useListHeight } from '../../hooks/useListHeight.js'
 import { TransactionHistoryEmpty } from './TransactionHistoryEmpty.js'
 import { TransactionHistoryItem } from './TransactionHistoryItem.js'
 import { TransactionHistoryItemSkeleton } from './TransactionHistorySkeleton.js'
@@ -19,9 +22,13 @@ export const TransactionHistoryPage: React.FC = () => {
   const { t } = useTranslation()
   useHeader(t('header.transactionHistory'))
 
+  const { listHeight } = useListHeight({
+    listParentRef: parentRef,
+  })
+
   const { getVirtualItems, getTotalSize } = useVirtualizer({
     count: transactions.length,
-    overscan: 10,
+    overscan: 5,
     paddingEnd: 12,
     getScrollElement: () => parentRef.current,
     estimateSize: () => 186,
@@ -35,36 +42,46 @@ export const TransactionHistoryPage: React.FC = () => {
 
   return (
     <PageContainer
-      ref={parentRef}
-      style={{ height: minTransactionListHeight, overflow: 'auto' }}
+      disableGutters
+      style={{ minHeight: minTransactionListHeight }}
     >
-      {isLoading ? (
-        <List disablePadding>
-          {Array.from({ length: 3 }).map((_, index) => (
-            <TransactionHistoryItemSkeleton key={index} />
-          ))}
-        </List>
-      ) : (
-        <List
-          style={{
-            height: getTotalSize(),
-            width: '100%',
-            position: 'relative',
-          }}
-          disablePadding
-        >
-          {getVirtualItems().map((item) => {
-            const transaction = transactions[item.index]
-            return (
-              <TransactionHistoryItem
-                key={item.key}
-                start={item.start}
-                transaction={transaction}
-              />
-            )
-          })}
-        </List>
-      )}
+      <Box
+        ref={parentRef}
+        style={{
+          height: listHeight,
+        }}
+        sx={{
+          overflow: 'auto',
+          paddingX: 3,
+        }}
+      >
+        {isLoading ? (
+          <List disablePadding>
+            {Array.from({ length: 3 }).map((_, index) => (
+              <TransactionHistoryItemSkeleton key={index} />
+            ))}
+          </List>
+        ) : (
+          <List
+            style={{
+              height: getTotalSize(),
+              position: 'relative',
+            }}
+            disablePadding
+          >
+            {getVirtualItems().map((item) => {
+              const transaction = transactions[item.index]
+              return (
+                <TransactionHistoryItem
+                  key={item.key}
+                  start={item.start}
+                  transaction={transaction}
+                />
+              )
+            })}
+          </List>
+        )}
+      </Box>
     </PageContainer>
   )
 }