+++ packages/widget/src/pages/TransactionPage/StartTransactionButton.tsx
@@ -1,6 +1,5 @@
 import { BaseTransactionButton } from '../../components/BaseTransactionButton/BaseTransactionButton.js'
-import { useFromTokenSufficiency } from '../../hooks/useFromTokenSufficiency.js'
-import { useGasSufficiency } from '../../hooks/useGasSufficiency.js'
+import { useMessageQueue } from '../../components/Messages/useMessageQueue.js'
 import type { StartTransactionButtonProps } from './types.js'
 
 export const StartTransactionButton: React.FC<StartTransactionButtonProps> = ({
@@ -9,21 +8,14 @@ export const StartTransactionButton: React.FC<StartTransactionButtonProps> = ({
   text,
   loading,
 }) => {
-  const { insufficientGas, isLoading: isGasSufficiencyLoading } =
-    useGasSufficiency(route)
-  const { insufficientFromToken, isLoading: isFromTokenSufficiencyLoading } =
-    useFromTokenSufficiency(route)
-
-  const shouldDisableButton = insufficientFromToken || !!insufficientGas?.length
+  const { hasMessages, isLoading } = useMessageQueue(route, true)
 
   return (
     <BaseTransactionButton
       onClick={onClick}
       text={text}
-      disabled={shouldDisableButton}
-      loading={
-        isFromTokenSufficiencyLoading || isGasSufficiencyLoading || loading
-      }
+      disabled={hasMessages}
+      loading={isLoading || loading}
     />
   )
 }